refactor: moved use-initializer to hooks in components
